Moved dirs to proper GoLang style and fixed README.md
