+++ analyzer/csv_exporter.go
@@ -1,8 +1,8 @@
 package analyzer
 
 import (
+	"github.com/VirtusLab/jira-stats/analyzer/domain"
 	"github.com/ztrue/tracerr"
-	"jira-stats/jira-stats-serverless/analyzer/domain"
 	"log"
 	"strconv"
 	"strings"
