+++ analyzer/db_access.go
@@ -2,13 +2,13 @@ package analyzer
 
 import (
 	"fmt"
+	"github.com/VirtusLab/jira-stats/analyzer/domain"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
 	"github.com/aws/aws-sdk-go/service/dynamodb/expression"
 	"github.com/ztrue/tracerr"
-	"jira-stats/jira-stats-serverless/analyzer/domain"
 	"log"
 	"time"
 )
