+++ analyzer/domain/domain.go
