+++ analyzer/domain/transition_interval_maker.go
