+++ analyzer/util.go
