+++ lambda_get/main.go
@@ -3,9 +3,9 @@ package main
 import (
 	"context"
 	"fmt"
+	"github.com/VirtusLab/jira-stats/analyzer"
+	"github.com/VirtusLab/jira-stats/analyzer/domain"
 	"github.com/ztrue/tracerr"
-	"jira-stats/jira-stats-serverless/analyzer"
-	"jira-stats/jira-stats-serverless/analyzer/domain"
 	"log"
 	"strings"
 	"time"
